fix(programs): consolidate lost validation into base PATCH, delete dead settings route (#477)
The dead PATCH /api/programs/[id]/settings route carried validation the
actually-used base PATCH /api/programs/[id] lacked. Port the missing
checks into the base handler and delete the dead route:

- Age bounds now use effective values (body-overrides-current) so a
  one-sided edit can't invert the minAge/maxAge range
- maxParticipants: positive-integer check + capacity floor against
  current enrollment count
- phase / enrollmentStatus: enum validation (rejects bad values with
  400 instead of letting Prisma throw a 500)
- announceOnOpen: boolean type check
- leadMentorId reassignment restricted to sysadmin/board only

Retarget the integration test from the deleted settings route to the
base PATCH handler — all assertions preserved.

Fixes #477
